PortFlow-Backend/portflow-agent/requirements.txt
django
django-ninja
python-decouple
ibm-watsonx-orchestrate
psycopg2-binary
uvicorn
requests
django-cors-headers
PyMuPDF
Pillow
pytesseract
python-multipart
PyJWT
